WIP: Migrating to Webpack 4
