+++ package.json
@@ -1,43 +1,33 @@
 {
   "name": "jspsych-react",
-  "version": "0.2.2",
-  "description": "A simple React component for jspsych",
-  "main": "dist/experiment.js",
+  "version": "0.3.0",
+  "description": "A React wrapper for jsPsych",
+  "main": "dist/jspsych-react.js",
   "repository": "git@github.com:makebrainwaves/jspsych-react.git",
-  "scripts": {
-    "build": "npm run clean && npm run build:dev",
-    "dist": "npm run clean && npm run build:prd",
-    "build:dev": "webpack src/experiment.js dist/experiment.js --config tools/webpack.config.dev.js",
-    "build:prd": "webpack src/experiment.js dist/experiment.js --config tools/webpack.config.prd.js",
-    "clean": "rimraf dist build",
-    "start": "node ./node_modules/webpack-dev-server/bin/webpack-dev-server.js"
-  },
+  "license": "MIT",
+  "private": false,
   "contributors": [
     "Teon L Brooks",
-    "Marion Le Borgne",
     "Dano Morrison",
+    "Marion Le Borgne",
     "William Wnekowicz"
   ],
-  "license": "ISC",
-  "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.4.0",
-    "babel-eslint": "^4.1.6",
-    "babel-istanbul": "^0.12.2",
-    "babel-istanbul-loader": "0.0.1",
-    "babel-loader": "^6.2.1",
-    "babel-preset-es2015": "^6.3.13",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-0": "^6.3.13",
-    "jquery": "^2.2.1",
-    "jsdom": "^8.1.0",
-    "react-addons-test-utils": "^0.14.7",
-    "rimraf": "^2.5.0",
-    "webpack": "^1.12.9",
-    "webpack-dev-server": "^1.14.0"
+  "scripts": {
+    "build": "npm run clean && npm run bundle",
+    "bundle": "webpack",
+    "clean": "rimraf dist build"
   },
   "dependencies": {
     "jspsych": "git://github.com/makebrainwaves/jspsych.git#jspsych-react@0.1.2",
     "react": "^0.14.3"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.5.0",
+    "@babel/preset-env": "^7.5.0",
+    "babel-loader": "^8.0.6",
+    "@babel/preset-react": "^7.0.0",
+    "rimraf": "^2.6.3",
+    "webpack": "^4.35.2",
+    "webpack-cli": "^3.3.5"
   }
 }
