+++ src/plugins.js
@@ -1,38 +1,35 @@
-import animation from "jspsych/plugins/jspsych-animation.js"
-import audio_button_response from "jspsych/plugins/jspsych-audio-button-response.js"
-import audio_keyboard_response from "jspsych/plugins/jspsych-audio-keyboard-response.js"
-import audio_slider_response from "jspsych/plugins/jspsych-audio-slider-response.js"
-import call_function from "jspsych/plugins/jspsych-call-function.js"
-import categorize_animation from "jspsych/plugins/jspsych-categorize-animation.js"
-import categorize_html from "jspsych/plugins/jspsych-categorize-html.js"
-import categorize_image from "jspsych/plugins/jspsych-categorize-image.js"
-import external_html from "jspsych/plugins/jspsych-external-html.js"
-import form from "jspsych/plugins/jspsych-form.js"
-import free_sort from "jspsych/plugins/jspsych-free-sort.js"
-import fullscreen from "jspsych/plugins/jspsych-fullscreen.js"
-import html_button_response from "jspsych/plugins/jspsych-html-button-response.js"
-import html_keyboard_response from "jspsych/plugins/jspsych-html-keyboard-response.js"
-import html_slider_response from "jspsych/plugins/jspsych-html-slider-response.js"
-import iat_html from "jspsych/plugins/jspsych-iat-html.js"
-import iat_image from "jspsych/plugins/jspsych-iat-image.js"
-import image_keyboard_response from "jspsych/plugins/jspsych-image-keyboard-response.js"
-import image_slider_response from "jspsych/plugins/jspsych-image-slider-response.js"
-import image_button_response from "jspsych/plugins/jspsych-image-button-response.js"
-import instructions from "jspsych/plugins/jspsych-instructions.js"
-import reconstruction from "jspsych/plugins/jspsych-reconstruction.js"
-import resize from "jspsych/plugins/jspsych-resize.js"
-import same_different_html from "jspsych/plugins/jspsych-same-different-html.js"
-import same_different_image from "jspsych/plugins/jspsych-same-different-image.js"
-import serial_reaction_time_mouse from "jspsych/plugins/jspsych-serial-reaction-time-mouse.js"
-import serial_reaction_time from "jspsych/plugins/jspsych-serial-reaction-time.js"
-import survey_likert from "jspsych/plugins/jspsych-survey-likert.js"
-import survey_multi_choice from "jspsych/plugins/jspsych-survey-multi-choice.js"
-import survey_multi_select from "jspsych/plugins/jspsych-survey-multi-select.js"
-import survey_text from "jspsych/plugins/jspsych-survey-text.js"
-import video from "jspsych/plugins/jspsych-video.js"
-import visual_search_circle from "jspsych/plugins/jspsych-visual-search-circle.js"
-import vsl_animate_occlusion from "jspsych/plugins/jspsych-vsl-animate-occlusion.js"
-import vsl_grid_scene from "jspsych/plugins/jspsych-vsl-grid-scene.js"
+import animation from "jspsych/plugins/jspsych-animation.js";
+import audio_button_response from "jspsych/plugins/jspsych-audio-button-response.js";
+import audio_keyboard_response from "jspsych/plugins/jspsych-audio-keyboard-response.js";
+import audio_slider_response from "jspsych/plugins/jspsych-audio-slider-response.js";
+import call_function from "jspsych/plugins/jspsych-call-function.js";
+import categorize_html from "jspsych/plugins/jspsych-categorize-html.js";
+import external_html from "jspsych/plugins/jspsych-external-html.js";
+import free_sort from "jspsych/plugins/jspsych-free-sort.js";
+import fullscreen from "jspsych/plugins/jspsych-fullscreen.js";
+import html_button_response from "jspsych/plugins/jspsych-html-button-response.js";
+import html_keyboard_response from "jspsych/plugins/jspsych-html-keyboard-response.js";
+import html_slider_response from "jspsych/plugins/jspsych-html-slider-response.js";
+import iat_html from "jspsych/plugins/jspsych-iat-html.js";
+import iat_image from "jspsych/plugins/jspsych-iat-image.js";
+import image_keyboard_response from "jspsych/plugins/jspsych-image-keyboard-response.js";
+import image_slider_response from "jspsych/plugins/jspsych-image-slider-response.js";
+import image_button_response from "jspsych/plugins/jspsych-image-button-response.js";
+import instructions from "jspsych/plugins/jspsych-instructions.js";
+import reconstruction from "jspsych/plugins/jspsych-reconstruction.js";
+import resize from "jspsych/plugins/jspsych-resize.js";
+import same_different_html from "jspsych/plugins/jspsych-same-different-html.js";
+import same_different_image from "jspsych/plugins/jspsych-same-different-image.js";
+import serial_reaction_time_mouse from "jspsych/plugins/jspsych-serial-reaction-time-mouse.js";
+import serial_reaction_time from "jspsych/plugins/jspsych-serial-reaction-time.js";
+import survey_likert from "jspsych/plugins/jspsych-survey-likert.js";
+import survey_multi_choice from "jspsych/plugins/jspsych-survey-multi-choice.js";
+import survey_multi_select from "jspsych/plugins/jspsych-survey-multi-select.js";
+import survey_text from "jspsych/plugins/jspsych-survey-text.js";
+import video from "jspsych/plugins/jspsych-video.js";
+import visual_search_circle from "jspsych/plugins/jspsych-visual-search-circle.js";
+import vsl_animate_occlusion from "jspsych/plugins/jspsych-vsl-animate-occlusion.js";
+import vsl_grid_scene from "jspsych/plugins/jspsych-vsl-grid-scene.js";
 
 const plugins = {
   animation,
@@ -42,7 +39,6 @@ const plugins = {
   call_function,
   categorize_html,
   external_html,
-  form,
   free_sort,
   fullscreen,
   html_button_response,
